Вынес extension и вспомогательные функции в отдельные файлы
